requirements.txt
tk
requests
Pillow
DrissionPage
